feat(Bridge): Update url
